+++ flutter_app/web/inline_script_4.js
@@ -1,0 +1,4 @@
+
+    import * as webllm from 'https://cdn.jsdelivr.net/npm/@mlc-ai/web-llm@0.2.84/+esm';
+    window.webllm = webllm;
+